refactor: 도메인 함수 분리
[#863]
